Trance-Checks/requirements.txt
pandas
numpy
spotipy
pillow